+++ templates/default/src/include/src/components/IndexHeader.vue
@@ -1,8 +1,10 @@
 <script setup>
-import { ref } from 'vue'
+import { ref, onMounted } from 'vue'
 
 import { useEmbedStore } from '../stores/embed'
 import { useFeedItemsStore } from '../stores/feedItems'
+import { useThemeStore } from '../stores/theme'
+import { useFiltersStore } from '../stores/filters'
 
 import IconGithub from './icons/IconGithub.vue'
 import IconHeart from './icons/IconHeart.vue'
@@ -20,7 +22,9 @@ const showScrollToTop = ref(false)
 const showRefresh = ref(false)
 const templateVersion = ref(window.templateVersion)
 
+const fStore = useFiltersStore()
 const embedStore = useEmbedStore()
+const themeStore = useThemeStore()
 const { resetFeedItems } = useFeedItemsStore()
 
 const props = defineProps({
@@ -29,6 +33,10 @@ const props = defineProps({
     required: true,
   },
 })
+
+const subHeaderText = ref(window.subHeaderText || null)
+const headerSvg = ref(window.headerSvg || null)
+const repoUrl = ref(window.repoUrl || null)
 
 const setScrollToTop = () => {
   let vh = Math.max(document.documentElement.clientHeight || 0, window.innerHeight || 0)
@@ -74,32 +82,59 @@ const toggleNav = () => {
   emit('toggle-nav')
 }
 
+const selectedTheme = ref(null)
+const selectTheme = () => {
+  themeStore.selectTheme(selectedTheme.value)
+}
+
 const refreshPage = () => {
   window.location.reload()
 }
+
+onMounted(() => {
+  selectedTheme.value = themeStore.themeName
+})
 </script>
 
 <template>
   <div class="header-crumbs">
     <span>
-      <h5>Page generated with <IconHeart /> by Liveboat</h5><br/>
-      <h5>Updated on {{ buildTime.toUTCString() }}</h5><br/>
+      <h5>
+        Page generated with <IconHeart /> by
+        <a href="https://github.com/exaroth/liveboat" target="_blank">Liveboat</a>
+      </h5>
+      <br />
+      <h5>Updated on {{ buildTime.toUTCString() }}</h5>
+      <br />
       <h5>Template ver. {{ templateVersion }}</h5>
     </span>
   </div>
-  <div class="header-container">
+  <div :class="{ 'header-container': true, 'header-container-archive': !props.feedList }">
     <div class="header-title">
       <h2>
-        <IconLiveboat />
+        <span id="header-custom-svg" v-if="headerSvg"><img :src="headerSvg" /></span>
+        <IconLiveboat v-else />
         <a :href="sitePath" v-html="pageTitle" />
       </h2>
+      <h3 v-if="subHeaderText" v-html="subHeaderText" id="header-subtitle" />
       <div id="icons-aggro">
-        <a id="icon-github" href="https://github.com/exaroth/liveboat" target="_blank">
+        <a
+          id="icon-github"
+          :href="repoUrl || 'https://github.com/exaroth/liveboat'"
+          target="_blank"
+        >
           <IconGithub />
         </a>
         <a id="icon-rss" href="rss.xml" target="_blank"><IconRss /></a>
         <a id="icon-opml" href="opml.xml" target="_blank"><IconOPML /></a>
       </div>
+    </div>
+    <div id="theme-selector">
+      <select v-model="selectedTheme" @change="selectTheme()">
+        <option v-for="(themeO, theme) in themeStore.availableThemes" :value="theme" :key="theme">
+          {{ themeO.txt }}
+        </option>
+      </select>
     </div>
   </div>
   <div id="side-buttons-wrapper">
@@ -114,7 +149,13 @@ const refreshPage = () => {
         @click="refreshPage()"
         ><IconRefresh
       /></a>
-      <a id="side-button-nav" v-if="props.feedList" title="Show navigation" @click="toggleNav()"><IconNav /></a>
+      <a
+        id="side-button-nav"
+        v-if="props.feedList && !fStore.firehose"
+        title="Show navigation"
+        @click="toggleNav()"
+        ><IconNav
+      /></a>
     </div>
   </div>
 </template>
@@ -123,8 +164,11 @@ const refreshPage = () => {
 .header-container {
   width: 100%;
   height: 80px;
-  margin: 10px 0px 10px 0px;
+  margin: 10px 0px 14px 0px;
   position: relative;
+}
+.header-container-archive {
+  margin: 10px 0px 50px 0px;
 }
 .header-title h2 {
   font-size: 1.8rem;
@@ -197,19 +241,19 @@ const refreshPage = () => {
 }
 
 #icons-aggro {
-  position: absolute;
+  position: relative;
   left: 0;
-  bottom: 0px;
   width: 200px;
+  z-index: 4;
 }
 
-#icon-rss svg,
-#icon-opml svg,
-#icon-github svg {
+#icon-rss,
+#icon-opml,
+#icon-github {
   fill: var(--color-text);
-  display: inline-block;
   float: left;
   margin-right: 10px;
+  display: block;
 }
 
 #icon-github svg {
@@ -223,6 +267,52 @@ const refreshPage = () => {
   height: 20px;
   top: 8px;
   left: 2px;
+}
+
+#theme-selector {
+  position: relative;
+  top: 4px;
+  z-index: 3;
+  width: 140px;
+  float: left;
+  border: 1px solid rgb(from var(--color-text) r g b / 40%);
+}
+#theme-selector select {
+  width: 100%;
+  height: 28px;
+  -webkit-appearance: none !important;
+  -moz-appearance: none !important;
+  appearance: none !important;
+  padding: 4px 8px;
+  outline: none;
+  border: none;
+  background-color: var(--color-background);
+  color: var(--color-text);
+}
+
+#theme-selector::after {
+  position: absolute;
+  content: '\2304';
+  top: -6px;
+  right: 8px;
+  color: rgb(from var(--color-text) r g b / 30%);
+  font-size: 1.4em;
+  pointer-events: none;
+}
+
+#header-subtitle {
+  display: box;
+  margin: 0 0 10px 0;
+  font-weight: 600;
+}
+
+#header-custom-svg img {
+  position: relative;
+  width: 36px;
+  height: 36px;
+  display: inline;
+  margin-right: 9px;
+  top: 4px;
 }
 
 @media (max-width: 640px) {
@@ -239,7 +329,6 @@ const refreshPage = () => {
   .header-crumbs {
     display: none;
   }
-
 }
 
 @media (max-width: 1640px) {
@@ -258,6 +347,13 @@ const refreshPage = () => {
 @media (max-width: 810px) {
   #side-button-nav {
     display: none;
+  }
+  #icons-aggro {
+    margin-top: 20px;
+    width: 100%;
+  }
+  .header-container-archive {
+    margin: 10px 0px 50px 0px;
   }
 }
 </style>