+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.1.4] 2025-02-03
+### Added
+- Add theme selector
+- Add support for `header-svg`, `repo-url` and `subheader-text` template settings (default template)
+### Fixed
+- Disable nav button when displaying firehose feed
+ 
 ## [1.1.3] 2025-01-27
 ### Fixed
 - Fix filtering for feeds with multiple tags (default template)